Flyttet aktivitetArbeidForBehandlingIds fra vurderingsservice til AndlesHistorikkService for å unngå loop. (Vil bruke AndlesHistorikkService i TidligereVedtaksperioderService. Da vil det bli loop via Vurderingsservice.
